Add sourcery binary
