+++ Tools/Sourcery/Templates/AutoCases.stencil
@@ -1,0 +1,10 @@
+{% for enum in types.implementing.AutoCases|enum %}
+{{ enum.accessLevel }} extension {{ enum.name }} {
+  static let count: Int = {{ enum.cases.count }}
+  {% if not enum.hasAssociatedValues %}
+  static let allCases: [{{ enum.name }}] = [
+  {% for case in enum.cases %}  .{{ case.name }}{{ ',' if not forloop.last }}
+  {% endfor %}]
+  {% endif %}
+}
+{% endfor %}
