+++ Tools/Sourcery/Templates/LinuxMain.stencil
@@ -1,0 +1,17 @@
+import XCTest
+
+{{ argument.testimports }}
+{% for type in types.classes|based:"XCTestCase" %}
+{% if not type.annotations.disableTests %}extension {{ type.name }} {
+  static var allTests: [(String, ({{ type.name }}) -> () throws -> Void)] = [
+  {% for method in type.methods %}{% if method.parameters.count == 0 and method.shortName|hasPrefix:"test" %}  ("{{ method.shortName }}", {{ method.shortName }}){{ ',' if not forloop.last }}
+  {% endif %}{% endfor %}]
+}
+{% endif %}{% endfor %}
+
+// swiftlint:disable trailing_comma
+XCTMain([
+{% for type in types.classes|based:"XCTestCase" %}{% if not type.annotations.disableTests %}  testCase({{ type.name }}.allTests),
+{% endif %}{% endfor %}])
+// swiftlint:enable trailing_comma
+
